remove references to pages for settings/teams
