+++ apps/web/app/future/settings/(settings)/teams/[id]/event-type/page.tsx
@@ -1,6 +1,7 @@
-import LegacyPage, { GetLayout } from "@pages/settings/teams/[id]/event-type";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
+
+import CreateTeamEventType, { GetLayout } from "~/settings/teams/[id]/event-types-view";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -8,4 +9,4 @@ export const generateMetadata = async () =>
     (t) => t("new_event_type_to_book_description")
   );
 
-export default WithLayout({ Page: LegacyPage, getLayout: GetLayout })<"P">;
+export default WithLayout({ Page: CreateTeamEventType, getLayout: GetLayout })<"P">;
